pyproject.toml
[project]
name = "procurement-discovery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.8.0",
    "black>=23.0.0",
    "fastapi>=0.100.0",
    "flake8>=6.0.0",
    "httpx>=0.24.0",
    "langchain>=0.1.0",
    "langchain-community>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.0",
    "markdown>=3.4.0",
    "pip>=25.2",
    "pydantic>=2.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.3.0",
    "uvicorn>=0.22.0",
]
